move all integration tests to new dsb-integration module; all tests compile now; WIP to modify them to pass
