--- dsb-gui/src/test/java/com/spectralogic/dsbrowser/gui/services/tasks/SearchJobTaskTest.java
@@ -1,85 +1,0 @@
-/*
- * ******************************************************************************
- *    Copyright 2016-2017 Spectra Logic Corporation. All Rights Reserved.
- *    Licensed under the Apache License, Version 2.0 (the "License"). You may not use
- *    this file except in compliance with the License. A copy of the License is located at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *    or in the "license" file accompanying this file.
- *    This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- *    CONDITIONS OF ANY KIND, either express or implied. See the License for the
- *    specific language governing permissions and limitations under the License.
- * ******************************************************************************
- */
-
-package com.spectralogic.dsbrowser.gui.services.tasks;
-
-import com.spectralogic.ds3client.commands.spectrads3.GetBucketsSpectraS3Request;
-import com.spectralogic.ds3client.commands.spectrads3.GetBucketsSpectraS3Response;
-import com.spectralogic.ds3client.models.Bucket;
-import com.spectralogic.dsbrowser.api.services.logging.LoggingService;
-import com.spectralogic.dsbrowser.gui.components.ds3panel.Ds3Common;
-import com.spectralogic.dsbrowser.gui.services.Workers;
-import com.spectralogic.dsbrowser.gui.services.newSessionService.SessionModelService;
-import com.spectralogic.dsbrowser.gui.services.savedSessionStore.SavedCredentials;
-import com.spectralogic.dsbrowser.gui.services.savedSessionStore.SavedSession;
-import com.spectralogic.dsbrowser.gui.services.sessionStore.Session;
-import com.spectralogic.dsbrowser.gui.util.ConfigProperties;
-import com.spectralogic.dsbrowser.gui.util.SessionConstants;
-import javafx.application.Platform;
-import javafx.embed.swing.JFXPanel;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mockito;
-
-import java.util.List;
-import java.util.Locale;
-import java.util.ResourceBundle;
-import java.util.concurrent.CountDownLatch;
-
-import static org.junit.Assert.assertTrue;
-
-public class SearchJobTaskTest {
-
-    private final Workers workers = new Workers();
-    private Session session;
-    private boolean successFlag = false;
-    private final static ResourceBundle resourceBundle = ResourceBundle.getBundle("lang", new Locale(ConfigProperties.getInstance().getLanguage()));
-
-    @Before
-    public void setUp() {
-        new JFXPanel();
-        Platform.runLater(() -> {
-            final SavedSession savedSession = new SavedSession(SessionConstants.SESSION_NAME,
-                    SessionConstants.SESSION_PATH, SessionConstants.PORT_NO, null,
-                    new SavedCredentials(SessionConstants.ACCESS_ID, SessionConstants.SECRET_KEY),
-                    false);
-            session = new CreateConnectionTask().createConnection(SessionModelService.setSessionModel(savedSession, false), resourceBundle);
-        });
-    }
-
-    @Test
-    public void search() throws Exception {
-        final CountDownLatch latch = new CountDownLatch(2);
-        Platform.runLater(() -> {
-            try {
-                final GetBucketsSpectraS3Request getBucketsSpectraS3Request = new GetBucketsSpectraS3Request();
-                final GetBucketsSpectraS3Response response = session.getClient().getBucketsSpectraS3(getBucketsSpectraS3Request);
-                final List<Bucket> buckets = response.getBucketListResult().getBuckets();
-                final SearchJobTask searchJobTask = new SearchJobTask(buckets, SessionConstants.TEXT_TO_SEARCH, session,
-                        workers, Mockito.mock(Ds3Common.class), Mockito.mock(LoggingService.class));
-                workers.execute(searchJobTask);
-                latch.countDown();
-                successFlag = searchJobTask.get() != null;
-                latch.countDown();
-            } catch (final Exception e) {
-                e.printStackTrace();
-                latch.countDown();
-                latch.countDown();
-            }
-        });
-        latch.await();
-        assertTrue(successFlag);
-    }
-}
