+++ dsb-gui/src/test/java/com/spectralogic/dsbrowser/gui/util/FileTreeTableProviderTest.java
@@ -21,7 +21,6 @@ import org.junit.Test;
 
 import java.io.File;
 import java.net.URL;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
